refactor(user): move user-related classes into dedicated package
