+++ tracker-api/src/main/java/com/sloyardms/trackerapi/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.sloyardms.trackerapi.repository;
 
-import com.sloyardms.trackerapi.entity.User;
+import com.sloyardms.trackerapi.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
