+++ tracker-api/src/main/java/com/sloyardms/trackerapi/service/TagServiceImpl.java
@@ -4,10 +4,9 @@ import com.sloyardms.trackerapi.dto.TagCreateDto;
 import com.sloyardms.trackerapi.dto.TagDto;
 import com.sloyardms.trackerapi.dto.TagUpdateDto;
 import com.sloyardms.trackerapi.entity.Tag;
-import com.sloyardms.trackerapi.entity.User;
+import com.sloyardms.trackerapi.user.entity.User;
 import com.sloyardms.trackerapi.exception.ResourceDuplicatedException;
 import com.sloyardms.trackerapi.exception.ResourceNotFoundException;
-import com.sloyardms.trackerapi.mapper.GroupMapper;
 import com.sloyardms.trackerapi.mapper.TagMapper;
 import com.sloyardms.trackerapi.repository.TagRepository;
 import com.sloyardms.trackerapi.repository.UserRepository;
