+++ tracker-api/src/main/java/com/sloyardms/trackerapi/user/dto/UserDto.java
@@ -1,4 +1,4 @@
-package com.sloyardms.trackerapi.dto;
+package com.sloyardms.trackerapi.user.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
