+++ tracker-api/src/test/java/com/sloyardms/trackerapi/controller/GroupControllerIntegrationTests.java
@@ -3,7 +3,7 @@ package com.sloyardms.trackerapi.controller;
 import com.sloyardms.trackerapi.dto.GroupCreateDto;
 import com.sloyardms.trackerapi.dto.GroupDto;
 import com.sloyardms.trackerapi.dto.GroupUpdateDto;
-import com.sloyardms.trackerapi.dto.UserCreateDto;
+import com.sloyardms.trackerapi.user.dto.UserCreateDto;
 import com.sloyardms.trackerapi.repository.GroupRepository;
 import com.sloyardms.trackerapi.security.DevSecurityConfig;
 import com.sloyardms.trackerapi.security.FakeAuthFilter;
