+++ tracker-api/src/test/java/com/sloyardms/trackerapi/controller/TagControllerIntegrationTests.java
@@ -3,7 +3,7 @@ package com.sloyardms.trackerapi.controller;
 import com.sloyardms.trackerapi.dto.TagCreateDto;
 import com.sloyardms.trackerapi.dto.TagDto;
 import com.sloyardms.trackerapi.dto.TagUpdateDto;
-import com.sloyardms.trackerapi.dto.UserCreateDto;
+import com.sloyardms.trackerapi.user.dto.UserCreateDto;
 import com.sloyardms.trackerapi.repository.TagRepository;
 import com.sloyardms.trackerapi.security.DevSecurityConfig;
 import com.sloyardms.trackerapi.security.FakeAuthFilter;
